+++ flytekit/core/context_manager.py
@@ -741,17 +741,6 @@ class FlyteContext(object):
 
         return _get_deck(typing.cast(ExecutionState, self.execution_state).user_space_params)
 
-    def add_deck(self, deck: Deck) -> None:
-        """
-        Adds a deck to the current context.
-        """
-        if (
-            self.execution_state is not None
-            and self.execution_state.user_space_params is not None
-            and hasattr(self.execution_state.user_space_params, "decks")
-        ):
-            self.execution_state.user_space_params.decks.append(deck)
-
     @dataclass
     class Builder(object):
         file_access: FileAccessProvider
