+++ flytekit/core/python_function_task.py
@@ -354,10 +354,14 @@ class PythonFunctionTask(PythonAutoContainerTask[T]):  # type: ignore
             source_code = inspect.getsource(self._task_function)
 
             from flytekit.deck import Deck
-            from flytekit.deck.renderer import SourceCodeRenderer
+            from flytekit.deck.renderer import PythonDependencyRenderer, SourceCodeRenderer
 
             source_code_deck = Deck("Source Code")
             renderer = SourceCodeRenderer()
             source_code_deck.append(renderer.to_html(source_code))
+
+            python_dependencies_deck = Deck("Python Dependencies")
+            renderer = PythonDependencyRenderer()
+            python_dependencies_deck.append(renderer.to_html())
 
         return super()._write_decks(native_inputs, native_outputs_as_map, ctx, new_user_params)
