+++ flytekit/deck/deck.py
@@ -174,67 +174,3 @@ def get_deck_template() -> "Template":
         autoescape=select_autoescape(enabled_extensions=("html",)),
     )
     return env.get_template("template.html")
-
-
-class PythonDependencyDeck(Deck):
-    """
-    PythonDependencyDeck is a deck that contains information about the library name and version.
-    """
-
-    def __init__(self, name: Optional[str] = "Python Dependencies", html: Optional[str] = ""):
-        super().__init__(name, html)
-
-    @property
-    def html(self) -> str:
-        import json
-        import subprocess
-        import sys
-
-        import pandas as pd
-
-        from flytekit.deck import MarkdownRenderer
-
-        try:
-            installed_packages = json.loads(
-                subprocess.check_output([sys.executable, "-m", "pip", "list", "--format", "json"])
-            )
-            requirements_txt = subprocess.check_output(["pip", "freeze"]).decode("utf-8").replace("\\n", "\n").rstrip()
-        except subprocess.CalledProcessError as e:
-            logger.error(f"Error occurred while fetching installed packages: {e}")
-            return ""
-
-        df = pd.DataFrame(installed_packages)
-        markdown = MarkdownRenderer().to_html(df.to_string(index=False))
-        html = f"""
-        <!DOCTYPE html>
-        <html lang="en">
-        <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Flyte Dependencies</title>
-        <script>
-        async function copyTable() {{
-          var requirements_txt = document.getElementById('requirements_txt');
-
-          try {{
-            await navigator.clipboard.writeText(requirements_txt.innerText);
-          }} catch (err) {{
-            console.log('Error accessing the clipboard: ' + err);
-          }}
-        }}
-        </script>
-        </head>
-        <body>
-
-        <button onclick="copyTable()">
-          <span>Copy table as requirements.txt</span>
-        </button>
-
-        {markdown}
-
-        <div id="requirements_txt" style="display:none">{requirements_txt}</div>
-
-        </body>
-        </html>
-        """
-        return html
